[doc] Update the pre-built-binary-distributions page
Remove an old broken link to ACL2s, update that section to link to the
ACL2s installation XDOC topics.

Add a section on Docker images, populate with a few of mine.

Add a link to the Homebrew package, and update the MacPorts reference to
link to the appropriate package.