Improved error messages when non-executable stobjs are present when evaluating a top-level form.  Updated notes about the "toothbrush".
Below is the update to notes about the "toothbrush" (file
~/acl2/acl2/books/system/toothbrush/README).  I discovered the failure
today when running on FreeBSD using CCL, and a few days ago Warren
Hunt reported a failure on a Mac using CCL.  I have been been doing
successful "toothbrush" runs on Linux with both CCL and SBCL.

NOTE added 8/2024: We have seen failures on FreeBSD using CCL that we
have not attempted to fix.  Other platforms may produce failures as
well.  As noted below, this is a work in progress; it should be viewed
as experimental.  By default, this directory is skipped during
regressions.  Only when environment TB_LISP is set will an attempt be
made to process anything in this directory.