[C$] Fix bug in parser.
When parsing a block item, the case of a labeled statement was not properly
accounted for. Now the case is accounted for, calling the (already existing)
code to parse the labeled statement as needed.

Thanks to Eric Smith for running the parser on an example randomly generated by
Csmith, which led to the discovery and fixing of this issue.