[C$] Turn parser state into "shallow" stobj.
Here "shallow" means that the `fty::defprod` has been turned into a stobj, but
its fields are the same, mostly lists, as the `fty::defprod` field, so those
lists are `cons`ed and such. But the parser state itself is updated
destructively.

Since the generated stobj operations do not fix their inputs and do not return
unconditionally typed outputs, we define our own fixtype-like operations that
wrap them. We also introduce a fixtype for the stobj itself. A few
read-over-write theorems (for the fixtype-like readers and writers, not the
stobj-generated ones) were also necessary.

Although the parser code was already written to treat the parser state in a
single-threaded way, this required a number of mostly systematic changes.

We plan to "deepen" the stobj, by turning the list fields into arrays and
indices.

Even with just this shallow stobj, on an example that was taking a while to
parse we get a 2x reduction in time and a 10x reduction in memory usage.

Tests and test macros also had to be adapted, in particular to use
`assert!-stobj`.