[AleoBFT dynamic] Refine model of committee calculation.
Enforce that the committee is always a subset of the total set of validators in
the system, by adding a parameter for that set to the operations that need it,
which is then instantiated in operations that have access to the system state.

This is actually more of a constraint on the system state than on the commitee
calculation proper. The documentation explains the rationale behind this
modeling approach.