Added macro attach-stobj and new defabsstobj keyword, :attachable, to conveniently arrange for an abstract stobj to have more than one foundation and associated functions for execution.  Fixed bug in handling of assert$ failures by break-on-error.  Removed a book from regression suite for GCL.  Tweaked :DOC defabsstobj.
Quoting :DOC note-8-6:

  A new [defabsstobj] keyword, :attachable, allows a single abstract
  [stobj] to have more than one foundation and associated functions
  for execution, without the need to recertify the book that
  introduces the stobj.  See attach-stobj.  Thanks to Warren Hunt and
  Yahya Sohail for requesting this feature.  We thank Sol Swords,
  Warren, and especially Yahya for helpful input on its high-level
  design.

To see the break-on-error bug, evaluate (break-on-error) followed by
(assert$ nil 17), to see a weird error involving fmt-abbrev not being
an embedded event form.  This bug was introduced after Version 8.5, so
there is no release note for it.

An attempt to certify
books/projects/x86isa/machine/interrupt-servicing.lisp resulted in an
error, as described in the following comment added there.  I'll
contact the book's author in case he wants to try to avoid this
problem.

; Matt K. addition: On an Intel-based 2019 MacBook Pro with 64 GB of memory,
; running "make -j 4 regression GCL_MEM_MULTIPLE=0.25" with ACL2 based on GCL,
; this book's certification failed with the following error (whitespace added
; here):
;   Error: Condition in CREATE-X86$C [or a callee]: INTERNAL-SIMPLE-ERROR: The
;   storage for RELOCATABLE-BLOCKS is exhausted. 161774 pages allocated. Use
;   ALLOCATE to expand the space.
; So I am excluding this book from regressions with ACL2 built on GCL.
; cert_param: (non-gcl)

In :DOC defabsstobj, not only mentioned the new :attachable keyword
argument, but also noted (this should have been there before) that
congruence requires not only matching up of the exports but also of
the recognizer and creator.

Successfully ran "make proofs" and "make devel-check" in addition to
the usual "make regression-fresh".