Make bigmem attachable and allow one to attach bigmem-asymmetric
This change is larger than one might naively expect because to allow for
attaching bigmem-asymmetric to bigmem, the logical definitions of the
stobjs must be the same (i.e. the same symbol). Previously, these
logical definitions, while identical, were repeated in each book under
their own package.

Thus, a centaur/bigmems/logic book was created with the logic
definitions. The bigmem(-asymmetric) directories were moved into that
directory as well for better organization and the doc topics were also
updated to reflect this hierarchy. The stobjs were also updated to use
the shared definitions in the logic book.

Thanks to Matt for implementing attachable stobjs.