Move reading from socket in tty-raw to separate thread
Warren observed very poor performance booting Linux on the x86isa model
when running under CCL on Mac OS with the tty-raw.lsp loaded. It seems
`read-char-no-hang,` which was previously called once per executed
instruction, is very slow on Mac OS.

To resolve this, we read the socket on a separate thread, which blocks
waiting for data. This thread places the read bytes in a queue (which is
a list for simplicity, but of course has the expected performance
implications) which can be quickly read from the main thread.

Thanks to Warren for reporting this issue.