[AleoBFT dynamic] Generalize non-equivocation.
This is non-equivocation of accepted certificates. We generalize it from a
single validator to multiple validators (with a single validator as a special
case). This requires assuming the blockchain non-forking invariant, from which
committee agreement follows, which we need for the quorum intersection arguments
(two of them) in the non-equivocation proof.