[AleoBFT dynamic] Slightly refactor some invariants.
Now that we have a dedicated invariant for validators being able to calculate
committees for accepted certificates, that conjunct has been removed from two
other invariant, which now rely on that invariant.

Some of the proof hints became more complicated, so it's not immediately clear
this was an overall improvement, but it still seems to factor the definitions
better, and perhaps we just need to add some targeted rules to bring the proofs
back to the previous complexity.