[AleoBFT stake] Add model of system states.
This is slightly simplified compared to the previous model with dynamic
committees and without stake: we only include correct validators in the map,
which should lead to further simplifications in other parts of the model. This
is actually orthogonal to whether we model stake or not, but it is an
opportunity to evaluate the advantages of this modeling simplification.