[C$] Fix bug in disambiguator.
Parentheses were occasionally dropped in `sizeof(E)` (with `E` an expression) or
some similar constructs. The reason is that in a `sizeof(X)` the `X` could be an
ambiguous expression or type name during parsing, recorded as such by the parser
and resolved after parsing by the disambiguator. But if `X` turns out to be an
expression, the actual operand of `sizeof` is `(E)`, not just `E`, because of
the way the C grammar is defined (it's asymmetric from `sizeof` applied to a
type name, which always requires parentheses). The same holds for some other
constructs with possibly ambiguous expressions or type names, but not all of
them. Thus we pass a flag to the disambiguation function that operates on
ambiguous expressions and type names, which, when set, instructs the function to
add parenthesese (in the AST) to the expression. The flag is set or cleared by
the caller of that function, depending on the construct.

Thanks to Eric Smith for running the tools on an example that revealed the issue
of dropped parentheses.