[C$] Extend results of mutually recursive validation functions.
Because of the GCC extension of statement expressions, the validation of every
mutually recursive construct can return, among its other results, a set of
possible return types. This commit extends all of these functions to do so, and
propagate and combine them accordingly.

We also complete the validation of expressions, with the remaining case of
statement expressions.

We also add the validation of some kinds of statements.