[C$] Fix bug in parser.
In some cases, parentheses around expressions were not preserved in the AST.
Although this did not cause incorrect behavior in semantic terms, the intention
of our parser is to preserve parentheses even when superfluous, and therefore
this is a bug.

Adapted some tests for a parsing function, whose arguments changed slightly (a
flag for parenthesization was added). Added a test to demonstrate preservation.

Thanks to Eric Smith for discovering and reporting the dropped parentheses in an
example.