[tools] Add `:extra-bindings-ok` to make-flag-generated macro lemma instantiation
When make-flag was invoked on a trivial mutual recursion with just one
function in the clique, the generated defthm-flag macro would fail on
subsequent theorems due to a generated lemma being instantiated with the
flag variable, but the lemma not actually containing said flag variable.
(The lemma would normally be a case statement on the flag variable, but
since the clique has just one member, the case statement only has the
trivial catch-all condition.) Adding the :extra-bindings-ok keyword
tells the system to just ignore the extra binding when it would
otherwise trigger an error.