[C$] Fix bug in parser.
In some cases, the parsing of an assembler statement neglected to parse the
required semicolon that concludes the statement.

Add a test demonstrating the fix.

Thanks to Eric Smith for running the parser on an example that revealed the
issue, which was manifesting as an extra null statement in the AST when printed
out after parsing, given that the ending semicolon was parsed separately from
the assembler statement. The newly added text checks that there is no extra
(null) statement.