[C$] Finish fixing a bug in the disambiguator.
This is the 3rd and hopefully last part of the fix. It turns out that the
adjustment of expressions w.r.t. priorities needs to performed recursively, to
cover all cases.

Thanks to Eric Smith for running the disambiguator on an example that uncovered
the need for this.

In the process, I also discovered that not only the construction of cast and
binary expressions, but also the construction of unary expressions needed
possibly adjustment based on relative priorities. This commit also handles unary
expressions.