[AleoBFT stake] Improve the model.
This eliminates the awkwardness of ignoring unbonding transactions that would
make the committee empty. Instead, we let committees become empty if
transactions work out that way, but we include a few extra checks in certain
state transitions to ensure that not only the stake of predecessor certificates
is at least the quorum, but also that it is positive. This ensures that there is
at least one predecessor, in every round greater than 1.

This improved model also lets us formally talk about what happens if committees
do become empty. It seems that the protocol would deadlock, but we plan to study
this in more detail in the future.