[axe/x86] Implement new scheme for clearing out shadowed writes.
This does not require resolving complex aliasing questions.  We push copies of clear-extend (an alias for clear) down into the write nest, interleaving them with the writes.  Once the shadowed write is reached, it is removed and the clear-extend is turned into a clear-retract (also an alias for clear).  The clear-retract moves up the nest, causing the clear-extends to be removed.  All this is done with rewrite rules.  This scheme does require an x86-specific axe-syntaxp function to recognize shadowed writes.  This is made possible by the user of the x86 variant of the Axe Rewriter in the x86 Lifter.