Make bigmem::mem and bigmem::mem$c non-executable
While bigmem::mem itself doesn't use much memory, if one attaches the
bigmem-asymmetric::mem to it, the memory usage can be substantial.
bigmem::mem is generally not used directly, so making it non-executable
to save on memory makes sense and if needed one can use
`add-global-stobj` to use it at the top level.

bigmem::mem$c is only used to back bigmem::mem so making it
non-executable as well shouldn't be an issue and saves some (although
probably insignificant) memory.