[RISC-V] Work around an XDOC issue.
Currently single-summand tagged sum fixtypes trigger a broader XDOC issue when
they are included in the manual. To work around this issue (which we should fix
at some point), we "manually" define the desired fixtype. The rest of the code
can still treat it as if it were defined in the desired way.