x86isa Fix large SSE2 shift counts causing Lisp errors
The SSE2 packed shift instructions which take the count as an XMM
register could have a huge (128-bit wide) shift count. These cause CCL
to throw when passed to `ash`.

Clamping the shift count to the el-width resolves this issue. This
commit also cleans up some of the SSE2 shift implementation code.