[AleoBFT] Simplify model.
Avoid notion of certificates associated to validators altogether, using instead
the notion of system certificates to define the notion of certificates signed by
a validator, which is the one used in the rest of the proofs. Also eliminate the
invariant that associated certificates are the same for all validators, as it is
no longer needed. This property is still proved in the `stake` version, so it
seems appropriate to simplify this `stake2` by eliminating that notion, along
with the other simplifications that characterize this `stake2` version.