[C$] Add `defpred` event macro.
This provides a way to concisely define predicates over the AST.

This commit adds the tool, along with user documentation.

We also include a test that shows how the macro can be used to define the
unambiguity predicates (whose current "manual" definition will be probably
eventually replaced by one generated by `defpred`).

Also extend and improve the top-level doc.

This `defpred` macro actually is close to becoming, after a few generalizations,
a generic tool for fixtypes.