[C$] Automate definition of unambiguity predicates.
Use `defpred` to concisely define the predicates. The generated predicates were
proved equivalent to the previously handwritten ones a few commits ago, so now
we just replace the handwritten ones with the generated ones, and we go forward
with them. This facilitates maintenance, as changes to the AST types (even just
simple renamings) are automatically adapted by `defpred`.

The rules for guard and return proofs are still handwritten, but we plan to
extend `defpred` to automate their generation as well.