[X86ISA] Reduce inclusion.
The book about initializing the x86 state doesn't need all the instruction semantic functions.  As is common with this sort of change, include-books had to be added to some downstream books, but overall the dependencies were reduced.

(This change was enabled by an earlier change that collected rules about XR.)