[C$] Simplify `input-files`.
Eliminates the `:read` option for the `:process` input, leaving only the
`:parse`, `:disambiguate`, and `:validate` options. There seems to be no real
point to just reading files in, and storing them into a `defconst` as a file
set. This also simplifies the result of the programmatic interface to be always
a translation unit ensemble, instead of either that or a file set.