[acl2s] Move ACL2s interface into a new dir
Moves ACL2s interface into a new directory (aspf) as we expect to
want to include other related code alongside it. This change also
includes stubs for the main books in the ACL2s interface so that
hopefully anyone using the library in the old location won't see
breakage before they can update to the new book location.