Fixed out-of-date info in :DOC deftheory (the example produces 60 runes, not 54).
Surely 54 was correct at some point in the past, but I like to avoid
anachronisms in the documentation that aren't acknowledged as such.
In my fix I point out that 60 is the number for Version 8.6; so I
won't need to fix that again.  Thanks to Warren Hunt for pointing this
out (a bit sheepishly, but actually I'm glad he did).