[vl/vl2014/misc] Fix error messages in SEQ books.
The Linter revealed that the calls to ER were missing the context arguments (so ER treated the format string as the context and the first value as the format 'string').  The calls of ER are in generated code, and it might be best to use the surrounding function name as the ctx.  However, I didn't see a nice way to access that, so I just used 'seq-process-bind as the context argument.