[RISC-V] Simplify some code.
Use a one-summand FTY sum type, now that it no longer causes issues in the
manual.

There is still a minor issue, namely that, when using the `case` macro for the
fixtype, the `.kind` variable is unused, which causes an error (so we add a
`set-ignore-ok t`). This issue could be overcome, by tweaking the FTY library to
generate an ignorable binding for that.