Modified state-global-let* to avoid call-arguments-limit problem in GCL 2.7.0.
Specifically, modified state-global-let*-fn to use list* instead of
list, to avoid a complaint by GCL 2.7.0 about exceeding the number of
arguments allowed for the function, list.  Thanks to Camm Maguire for
his help in locating this problem.

Also, added a comment in complile-acl2 about a small issue with GCL
2.7.0 that is probably not a serious problem, to which I may return.