[FTY] Tweak `deffold-reduce`.
For lists and maps, avoid the special treatment of singletons. The special
treatment was motivated by uniformity with the treatment of product types, but
it makes more sense for lists to follow the normal form of list folds, and for
omaps, which are structurally similar, to do that as well.

Thanks to Grant Jurgensen for suggesting this change.