[C2C] Fix call graph check on direct function calls
Previously, the call graph generation naively assumed function calls to
be direct if the function is just an identifier. This is of course not
the case in general; we can have a call of `foo(...)` where `foo` is not
the name of a function, but a function pointer. We now check the linkage
of `foo` to ensure it is external or internal, and that it is bound a
function type object.

This also adds a test where the previous implementation failed.