[axe] Clarify rewriter function names.
Most generated rewriter functions, such as simplify-dag-xxx and simplfy-term-xxx return DAGs.  Those that return terms now have names that indicate that.  So now we have simplify-term-to-term-XXX (previously simp-term-XXX), and similarly simplify-terms-to-terms-XXX (was simp-terms-XXX).  Recall that the XXX stands for the suffix of the specific rewriter being generated, such as rewriter-basic or rewriter-x86.