Include virtually all the new set-theory books in books/top.lisp.  Tweak :doc int=.
I noticed that books/top.lisp has the form
(include-book "projects/doc" :dir :system)
and that, with only one exception, this is the only inclusion of
projects/ books in books/top.lisp.  I think the idea of
books/top.lisp, as opposed to books/doc/top.lisp, is to include most
of the community books, to help maintain avoidance of name conflicts
(including package conflicts).  Now books/projects/doc is based on
books/projects/include-doc.lisp, where I recently added
(include-book "set-theory/doc").  Here I am changing that to
(include-book "set-theory/top")
so as to get virtually all the set-theory books included with
books/top, not merely the documentation.
In fact this commit includes a small change to one event in the
set-theory books to avoid a name conflict.

The tweak to :DOC int= is due to a report from Grant Jurgensen
(thanks, Grant!) that he observed = running faster on integers than
int= does, for the CCL version he was using.  (I view this as a
deficiency of that Lisp.)