[axe] Improve the handling of :types in the main (legacy) proof tools.
Now the :types argument only allows axe-types, whereas more exotic types for test-based sweeping-and-merging (like :list and :range types) must be given using the new :test-types argument.  Only the :types are assumed true by the proofs, not the :test-types.  The :test-types are effectively just hints to help in the testing process.