[C2C] Extend `simpadd0`.
Also simplify `E + 0` to `E` when `E` is an expression that the validator
annotates as having type `int`, without requiring it to be a variable.

Also fix a bug in proof generation for functions with multiple parameters, which
is independent from the above extension but came up when testing that extension.