Proved Cantor's theorem.  Did a bit of clean-up in set theory books.
Quoting a comment in books/projects/set-theory/cantor.lisp,
which has Cantor's Theorem:

; This proof closely follows the proof of Cantor's theorem as presented in the
; Wikipedia article of that name (as of 4/5/2025), at the following URL.

; https://en.wikipedia.org/wiki/Cantor%27s_theorem

I'll let this settle a bit before asking for an update to the
"Formalizing 100 Theorems" page (https://www.cs.ru.nl/~freek/100/).
Thanks to Grant Jurgensen for pointing out that Cantor's theorem is
on that page without an ACL2 proof mentioned there.

The main change to the set theory books was to disable the rule
in-natp at its introduction in base.lisp, to avoid having to disable
it when proving Cantor's theorem.  The other changes were not
necessary for the proof of Cantor's theorem.  They included: remove a
couple of disables of in-natp that are no longer necessary and add a
theory-invariant involving in-natp; strengthen the one existing
theory-invariant to cause an error instead of a warning; add comments
in base.lisp and rename axiom cons-as-pair to cons-as-ordered-pair,
for clarity; and make tiny tweaks to the documentation.

With Cantor's theorem proved, I'll probably next deal with
cardinality.