[C$] Add optional tag to struct types
The interpretation of (type-struct (ident "foo")) is the collective type
of all struct types with tag "foo". To be perfectly precise, one would
need to distinguish structs types further based on the scope of their
declaration. (Of course, we'd also like to eventually store the struct
members -- their order, names, and types).

Similarly, (type-struct nil) is the collective type of all tagless
structs. To be more precise in this case, we'd need to also represent
the location of its declaration, as each declaration of a tagless
structure type denotes a new type [C17:6.7.2.3/5].

The validator has been updated to recognize different structure types
and check for compatibility as appropriate.