Translated the ACL2 publications page into xdoc and included the result in the manual.
There seems to be a sense in the ACL2 community that it would be good
to move more ACL2-related documentation into the (xdoc) manual.  To
that end, this commit introduces a new book,
books/doc/publications.lisp, which contains an xdoc version of
https://www.cs.utexas.edu/~moore/publications/acl2-papers.html.  The
name of the top-level topic for the new publication topics is
publications (in the "ACL2" package and imported into the new
"PUBLICATIONS" package).

My plan is to change the original web page acl2-papers.html (mentioned
above) to contain nothing more than a link to the new xdoc topic.  I
may do this in about a month.  In the meantime, that page can of
course be used as a reference to check the translation.  Note that the
translation of that single page is into many :DOC topics, in part
because xdoc does not support <a name=...> or <a href=...#...>.

Of course, translation errors should be fixed, and if they are
reported to me then I'd be happy to deal with them.

Please note that the original acl2-papers.html has been only lightly
maintained over time -- it is missing many suitable papers and it
could benefit from some updating.  I corrected only very few such
issues when I did the translation to xdoc.  Please do not ask me to
make changes other than corrections to translation errors or to
entries based on my own papers -- it's now up to the community to
maintain this information in xdoc.

As noted in :DOC publications, the workshops proceedings "contain
numerous papers, many of them recent, that are not found".  I'll leave
it to the community to decide whether or not this comment somewhat
covers missing workshop papers, so that missing non-workshop papers
have the highest priority.