Significantly improved the statement of the Schroeder-Bernstein Theorem (SBT) in set theory.  This required an ACL2 enhancement to allow "strict" table guards, which restrict table updates to avoid using the variables WORLD or ENS.  For the SBT, the new approach makes the zfc-table guard strict, but weakens its check so that in the second pass of an encapsulate for which the proposed table key is in the signatures, the value check is skipped.  The new file books/projects/set-theory/logic.txt contains statements specifying the foundations for the ACL2 set-theory work, with proofs.  Also in support of the improvement is that extend-zfc-table creates a function, not a macro, and force? is now just force.  Also, fixed a bug unduly restricting the use of ENS in table key expressions for updates.
Note: The new logic.txt file essentially supersedes the Section on
"Interpreting ACL2 theories in set theory" in :DOC zfc-model.  I may
rework that :DOC topic to account for the much more complete
development in logic.txt.  Moreover, I might mark up logic.txt into a
:DOC topic, perhaps incorporating it into :DOC zfc-model; if anyone
particularly wants me to do so, please contact me to discuss that.

(Aside: By far the majority of this work was in developing logic.txt.
It would be interesting to know if anyone reads it, or attempts to do
so.)

More documentation may be forthcoming, in particular for the new
macro, defthmz+, used in the statement of the SBT.

Quoting :DOC note-8-7:

  The new function symbol strict-table-guard returns its single
  argument unchanged, but has a special meaning when a [table]'s
  :guard is a call of that function symbol.  In that case, a term
  supplied for updating the table --- that is, for the :put and
  :clear operations --- must have no free variables, even though
  variables WORLD and ENS are normally permitted.  See [table].

  According to the documentation for [table], the variable ENS is
  allowed in both the key and value expressions when updating the
  table, that is, using the :put and :clear operations.  However, ENS
  was not being allowed in the key expression.  That has been fixed.