[AleoBFT] Improve model of blocks.
Build round evenness into fixtype of blocks, as done in the next version of the
formal model and proofs. Adapt proofs as needed. Still need to simplify some
predicates about both ordered and even (for blocks and certificates) to be just
ordered, as done in the next version of the formal model and proofs.