[IHS] Add two fixing theorems.
These are disabled because, if enabled, they cause failures elsewhere.

These are useful when `ifix` is kept disabled, which is perhaps uncommon, but
avoids, like other built-ins, unnecessary case splits.

We could consider enabling these theorems and just repair the failing proofs (by
disabling them, or one of them), but this is a bigger change.