[axe/x86] Implement new scheme for 32-bit registers.
Now, for 32-bit proofs, the register getters and setters in our normal form deal only with the low 32-bits of each register.  Because the 32-bit instructions can affect bits above these (due to code reuse from the 64-bit instructions), we introduce machinery to reason about the high half of the registers.  The 32-bit instructions seem to either preserve these bits or set them to 0, so we add assumptions that these bits are initially 0 (in 32-bit mode, where they are irrelevant).  These new assumptions allow rewrite rules to simplify away the functions affecting the high halves of the registers.