[APT] Extend `isodata`.
Allow input and output stobjs. If their representation is transformed, the
generated function is non-executable (i.e. `defun-nx`), since the application of
the conversions does not satisfy the stobj restrictions. See added user
documentation in regard.

Remove tests for rejecting stobjs.

An attempt to add tests involving stobjs generates errors in the `defun-nx`, not
in the body of the function, but in its guard. It is not clear whether this is
an issue in ACL2 or not; if it is, we will add `isodata` tests once the issue is
resolved; if it is not, we will have to generate a more elaborate guard that
works around the restriction.