Provided a workaround for allowing spaces in pathnames; modified :DOC accordingly.
Quoting :DOC note-8-7:

  By default, the directory where an ACL2 executable is to be built
  must not have a pathname that contains spaces, as before.  However,
  now there is a variable, ACL2_ALLOW_SPACES_IN_DIRECTORIES, that may
  be set to a non-empty value in order to build an ACL2 executable in
  such a directory, as noted in the error message.  That message
  points out that there may be errors, however, when certifying
  books.  Thanks to Eric Smith for a discussion leading to this
  change.

More information:

Eric Smith has pointed out that it has put off an important potential
user that ACL2 could not be installed under a directory whose pathname
contains a space.  It turns out that even with such a directory, ACL2
can be built in five of the six supported Lisps, and certify-book can
be used successfully in those cases.  However, the build system for
books would need modification to support the use of cert.pl or (for
example) "make regression".

If Eric and the community want to work on modifying the build system
for books to support spaces in the community books directory, they now
have a means to experiment with such modifications, by setting an
environment or "make" variable as noted below.  That variable is now
used in both ./GNUmakefile and books/GNUmakefile to avoid immediate
errors from spaces in pathnames (but again, the build system for books
will need fixing to avoid such errors entirely).

Yet more details:

The error message when attempting such a build is now preceded by a
message such as the following (with a suitable pathname on the first
line).

- Illegal ACL2 build directory (contains a space): /u/jones/spaces test/
-   To avoid this error, try setting variable ACL2_ALLOW_SPACES_IN_DIRECTORIES
-   to a non-empty value, e.g., (make ACL2_ALLOW_SPACES_IN_DIRECTORIES=t).
-   But although that may allow you to build saved_acl2, there may be
-   errors when attempting to certify books; so use this workaround at
-   your own risk.

Since this information is provided at build time, I modified :DOC
installation-summary by removing the words about installing ACL2 under
a directory "whose full pathname contains no whitespace".  It seemed
to Eric and me that it's not necessary to clutter the installation
instructions with that caveat; few will try with spaces in pathnames,
and those who do will get the message above.

Quoting a comment from ./GNUmakefile:

# The build of saved_acl2 may succeed even if the directory name
# includes spaces; we have seen this success using each supported Lisp
# except GCL (Version 2.7.1).  However, book certification with the
# build system for books, such as with "make" or with
# books/build/cert.pl, will fail (at least, as of this writing in July
# 2025); so we disallow such a build by default.  The message below
# provides a workaround for building saved_acl2 but not for the issue
# with the build system for books.