[doc] Remove obtaining-acl2 topic
This was essentially the same procedure as in installation-instructions.  Some minutiae about tar and WinZIP have been moved to windows-installation.  I did remove the discussion of disk space used by ACL2.  That seemed unnecessary in 2025 and likely to become out-of-date.  And it didn't include the only important number: the amount of space taken after certifying books.