[C2C] Fix bug in split-fn
This fixes a bug where nullary C functions were not split. This was
because the declarator in the AST for the function definition uses a
different constructor than regular function definitions which was not
being handled.

Also adds a test case of a nullary function.

Thanks to Eric Smith for discovering and reporting this bug.