Slightly relaxed the new restrictions on function attachment pertaining to abstract stobjs.  Tweaked some Emacs code.
Thanks to Sol Swords for suggesting this relaxation, which restricts
consideration of supports of abstract stobj primitives to those that
return the new stobj.

Here is a bit of technical detail.  I replaced function
absstobj-methods-logic-and-exec by absstobj-producer-fns, which has
the following comment.  (It references "Part 3" of an essay that is
not yet updated.)

; Part 3 of the Essay on the Correctness of Abstract Stobjs explains how the
; canonical correspondence function is defined from the creator and exports.
; Only the exports that can actually change the stobj can contribute to that
; function's definition.

The Emacs code tweaks were made in response to warnings from the byte
compiler for Emacs 30.  Thanks to Warren Hunt for bringing these
warnings to my attention.