Improve efficiency of web manual search
The shorts are no longer being tokenized. Also, when we search the
shorts, we use their exact (XML) contents. Searching the "rendered"
version is much too slow. These two changes make searching very fast.

Also removes highlighting of the search term on navigating to a topic.
This is a compelling feature, but it requires more careful
consideration. We should probably highlight more than just the first
instance of the search term in the page. And it should be flexible in
the same way the main search is (tokenize the query, allow partial match
of tokens).