[C$] Fix bug in parser.
The handling of what follows a `sizeof (` was not complete in all cases. We fix
the issue in a systematic way: based on the grammar, after parsing `sizeof (`,
we put back the `(` and then we attempt to parse a possibly ambiguous unary
expression (which in fact must be a postfix expression, since it starts with
`(`) or parenthesized type name. This is, in a way, a specialization of the more
general function to parse a possibly ambiguous expression or type name. The
difficulty, here, is caused by the fact that, when `sizeof` is used as a unary
expression, as opposed as to an operator on type names, the argument expression
needs not be parenthesized. Although the parser was already aware of that, the
problem was that something like `(x)` may be the start of a longer postfix
expression such as `(x).m` or `(x)->m`.

Thanks to Grant Jurgensen for running the parser on an example that led to the
discovery of this bug.