Modified change made yesterday for inlining to apply only to SBCL, not Allegro CL or CMUCL.
For Allegro CL and CMUCL, the change that was made yesterday was
leaving inlined functions from a book as uncompiled when the book is
included.  It might be easy enough to make the change properly for
those other two Lisps as well, but I'm not currently interested in
spending time on that since those Lisps rarely host ACL2.  So for
those Lisps, an inlined function from a book will run compiled but
might not be inlined.