Considerably improved the Essay on the Correctness of Abstract Stobjs.
It remains to extend the main proof to cover the case that an abstract
stobj has a foundation that is an abstract stobj, as well as the case
that a child stobj is an abstract stobj.  But this has been a taxing
effort, so I plan to wait awhile before taking on that extension.