[C$] Fix handling of unary `&&` GCC extension.
Despite the fact that unary `&&` may be only followed by an identifier, we need
to generalize the AST to accommodate the possibility of an inclusive disjunction
expression following an ambiguous `&&`. We also need to treat its disambiguation
in the same way as similar unary/binary ambiguities, with slight variations due
to the aforementioned restriction on identifiers.

This commit also fixes an unrelated possible mishandling of the
increment/decrement operators that may intervene in this and other kinds of
ambiguous expressions.

Thanks to Grant Jurgensen for running the C parser on an example that revealed
this issue. The fixes to the disambiguator are related, although the test was
failing just on the parsing.

This commit also adds the example as a parser test and as a disambiguator test.