[C$] Improve AST nomenclature.
Rename 'blocks' to 'compound statements' (fixtype `comp-stmt`), in line with the
concrete syntax. Indeed, according to the C standard, blocks are not quite
synonyms of compound statements: the latter are sequences of block items
enclosed in braces, which are also blocks, but there are other things that are
blocks, without being compound statements; examples are loop statements and
their bodies.