Fixed ACL2(p) bug that was recently introduced.  Removed waterfall-parallelism from a few books to prevent ACL2(p) certification failures.
My ACL2(p) regression caused a lot of errors shown in .cert.out files
as "Exit code from ACL2 is 0".  That's become common, and ACL2(p)
testing is no longer supported (e.g., for releases).  In fact it was
so common in this run that I terminated it early.  But I found some
books with explicit failures, and I turned off waterfall-parallelism
in those.

It might be awhile before ACL2(p) is tested again, other than to build
it.