[C] Slightly improve dynamic semantics.
In preparation for having `exec-expr` handle the right side of an assignment
directly (i.e. without resorting to `exec-expr-pure`), we add a check, in
`exec-expr`, that the right side is pure unless the left side is a variable.

This simple added check turned out to require quite a few proof adaptations, in
the form of `:expand` hints, because apparently the slight increase in the
complexity of the body of the dynamic execution function trigger some heuristics
in which the definition is not expanded as it should, and so we need to force
the expansion.