[C2C] Eliminate use of pure expression execution.
Now all the generated theorems are based on `exec-expr` instead of
`exec-expr-pure`. This is a larger commit than usual because there did not seem
to be an easy way to do the change more gradually. It is based on the extension
of the dynamic semantics to handle non-strict expressions (i.e. `&&`, `||`, and
`?:`) in `exec-expr`.