[ATC] Refine some generated proofs.
In certain theorems, just enabling the guard function is not enough to control
the proofs, because of the normalization of definitions. So, after the event
that defines each guard function, we also generate an event that installs a
normalized definition, and we use that in generated hints.

An alternative could be to directly define the guard functions as not
normalized, since they are only used internally anyhow for now. But in the
future we may want to export them, so this approach is more useful.