[ATC] Consolidate some generated proofs.
For a loop function, instead of generating a variant of the termination theorem
and then using it to generate the measure theorem, we generate directly the
latter, making use of ACL2's internal termination theorem directly. It is the
measure theorem that is used in the recently optimized loop proofs.